Move CharCode from kotlin.text to kotlin package
