+++ libraries/stdlib/js-ir/builtins/Char.kt
@@ -12,7 +12,10 @@ package kotlin
 // TODO: KT-35100
 //@Suppress("NON_PUBLIC_PRIMARY_CONSTRUCTOR_OF_INLINE_CLASS")
 //public inline class Char internal constructor (val value: Int) : Comparable<Char> {
-public class Char internal constructor(private val value: Int) : Comparable<Char> {
+@Suppress("ACTUAL_WITHOUT_EXPECT", "CONFLICTING_OVERLOADS")
+@OptIn(ExperimentalUnsignedTypes::class)
+public class Char actual constructor(code: UShort) : Comparable<Char> {
+    private val value: Int = code.toInt()
 
     /**
      * Compares this value with the specified value for order.
