+++ libraries/stdlib/src/kotlin/text/Char.kt
@@ -8,50 +8,6 @@
 
 package kotlin.text
 
-
-/**
- * Creates a Char with the specified [code], or throws an exception if the [code] is out of `Char.MIN_VALUE.code..Char.MAX_VALUE.code`.
- *
- * If the program that calls this function is written in a way that only valid [code] is passed as the argument,
- * using the overload of this function that takes a [UShort] argument is preferable (`Char(intValue.toUShort())`).
- * That overload doesn't check validity of the argument, and may improve program performance when the function is called routinely inside a loop.
- *
- * @sample samples.text.Chars.charFromCode
- */
-@ExperimentalStdlibApi
-@SinceKotlin("1.4")
-@kotlin.internal.LowPriorityInOverloadResolution // to not clash with Char constructor in js-ir
-public fun Char(code: Int): Char {
-    if (code < Char.MIN_VALUE.code || code > Char.MAX_VALUE.code) {
-        throw IllegalArgumentException("Invalid Char code: $code")
-    }
-    return code.toChar()
-}
-
-/**
- * Creates a Char with the specified [code].
- *
- * @sample samples.text.Chars.charFromCode
- */
-@OptIn(ExperimentalUnsignedTypes::class)
-@ExperimentalStdlibApi
-@SinceKotlin("1.4")
-public fun Char(code: UShort): Char {
-    return code.toInt().toChar()
-}
-
-/**
- * Returns the code of this Char.
- *
- * Code of a Char is the value it was constructed with, and the UTF-16 code unit corresponding to this Char.
- *
- * @sample samples.text.Chars.code
- */
-@ExperimentalStdlibApi
-@SinceKotlin("1.4")
-@kotlin.internal.InlineOnly
-public inline val Char.code: Int get() = this.toInt()
-
 /**
  * Returns the numeric value of the decimal digit that this Char represents.
  * Throws an exception if this Char is not a valid decimal digit.
